+++ app/src/main/java/woowacourse/shopping/data/db/cartItem/CartItemConverters.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.db
+package woowacourse.shopping.data.db.cartItem
 
 import androidx.room.TypeConverter
 import com.google.gson.Gson
