+++ app/src/main/java/woowacourse/shopping/data/model/CartItemEntity.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.data.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import woowacourse.shopping.data.db.CartItemDatabase.Companion.CART_ITEMS_DB_NAME
+import woowacourse.shopping.data.db.cartItem.CartItemDatabase.Companion.CART_ITEMS_DB_NAME
 import woowacourse.shopping.domain.Product
 
 @Entity(tableName = CART_ITEMS_DB_NAME)
